Implement message sending function
1. Set index action and create actions in messages controller
2. Define “set_group” action under private and set as before_action in
the messages controller. This makes it possible to use @group for all
actions in the messages controller.
In create action, make conditional branch for both case of success to
save and not.
3. Edit view file. Use form_for method. Put @group and @message in the
argument box. Since messages are nested below the groups, form_for
@message is not going to work. the correct path to create action is
group_messages_path(not messages_path, which is created if only put
@message in the form_for argument box.