Divide processing with respond_to method
with respond_to method, set the conditional branches that is when the
receiving request is html, redirect_to messages#index, and the
receiving request is json, jbuilder is converting the input data into
json format.
In the create.json.jbuilder file define values in key-value format that
is possible to use in the JavaScript file.