Automatic update
1. message.js
Set the variance of “interval” by using  setInterval method, and read
the autoUpdate function every 5 seconds unless the user stays at the
same chat page.

2. message.js / _message.html.haml
make function autoUpdate
- if class “messages” has messages, send the very last message’s id
to the controller as json datatype, and if there is no messages, the id
“0” is sent.

3. messages_controller.rb
- with respond_to divide the process by the data type html or json.
- if the json format, when the newer id than the id of current
viewing chat page’s very last message does not exist on the database,
controller deliver the data of the @new_messages.

4. index.json.jbuilder
- json data received in the controller is sent throught jbuilder, and
define those data in the jbuilder for use in message.js file.

5. message.js
- receive the data from jbuilder
- if success, the new messages are inserted to the very last place of
“.messages”.
- if fail, show alert